+++ services/cases_other.js
@@ -1,11 +1,15 @@
 const Case = require('../models/Case')
 const History = require('../models/History')
 const ObjectId = require('mongodb').ObjectID
-const { CRITERIA, WHERE_GLOBAL, HISTORY_DEFAULT_SORT, ROLE } = require('../helpers/constant')
+const {
+  CRITERIA, WHERE_GLOBAL,
+  HISTORY_DEFAULT_SORT, ROLE
+} = require('../helpers/constant')
 const { clientConfig } = require('../config/redis')
 const { thisUnitCaseAuthors } = require('../helpers/cases/global')
 const { countByRole } = require('../helpers/rolecheck')
 const { filterCase } = require('../helpers/filter/casefilter')
+const { summaryCondition } = require('../helpers/cases/global')
 
 const conditional = async (result, payload, val) => {
   if (result.length > 0) {
@@ -102,13 +106,84 @@ async function getCaseSummary(query, user, callback) {
   }
 }
 
+async function healthCheck(payload, callback) {
+  try {
+    let case_no_last_history = await Case.find({ last_history: {"$exists": false}})
+    .or({ last_history:null })
+    let result = {
+      'case_no_last_history' : case_no_last_history,
+    }
+    return callback(null, result);
+  } catch (error) {
+    return callback(error, null)
+  }
+}
+
+async function getCaseSummaryVerification (query, user, callback) {
+  // Temporary calculation method for faskes as long as the user unit has not been mapped, todo: using lookup
+  let condition
+  if (user.unit_id) {
+    condition = { unit_id: user.unit_id._id, role: ROLE.FASKES}
+  } else {
+    condition = { unit_id: user.unit_id._id }
+  }
+  const caseAuthors = await thisUnitCaseAuthors(user, condition)
+  const searchByRole = countByRole(user,caseAuthors);
+  const filterSearch = await filterCase(user, query)
+  const searching = {...searchByRole, ...filterSearch}
+  var aggStatus = [{ $match: {
+      $and: [  searching, { delete_status: { $ne: 'deleted' } } ]
+    }},
+    { $group: { _id: "$verified_status", total: { $sum: 1 }} }
+  ];
+
+  let result =  { 'PENDING': 0, 'DECLINED': 0, 'VERIFIED': 0 }
+  Case.aggregate(aggStatus).exec().then(async item => {
+      item.forEach(function(item){
+        summaryCondition(item, 'pending', result, 'PENDING')
+        summaryCondition(item, 'declined', result, 'DECLINED')
+        summaryCondition(item, 'verified', result, 'VERIFIED')
+      })
+      return callback(null, result)
+    })
+    .catch(err => callback(err, null))
+}
+
+function getCountCaseByDistrict(callback) {
+  const aggStatus = [
+    { $match: { delete_status: { $ne: 'deleted' }} },
+    {$group: {
+      _id: "$address_district_name",
+      total: {$sum: 1}
+    }}
+  ];
+
+  const result =  {}
+
+  Case.aggregate(aggStatus).exec().then(item => {
+      item.forEach(function(item){
+        result[item['_id']] = item['total']
+      });
+      return callback(null, result)
+    })
+    .catch(err => callback(err, null))
+}
+
 module.exports = [
   {
     name: 'services.cases_other.multipleUpdate',
     method: multipleUpdate
-  },
-  {
+  },{
     name: 'services.cases_other.getSummary',
     method: getCaseSummary
+  }, {
+    name: 'services.cases_other.getSummaryVerification',
+    method: getCaseSummaryVerification
+  }, {
+    name: 'services.cases_other.healthcheck',
+    method: healthCheck,
+  }, {
+    name: 'services.cases_other.getSummaryByDistrict',
+    method: getCountCaseByDistrict
   },
 ]
